requirements.txt
Pillow

pytest
codecov
pytest-cov
